Cleaned up the tests locally because I can know run and dev locally :D. Also moved the deepspeed files to their own folder.
